Moved FireStoreClient to app module. Changed its implementation using generics. Moved FirebaseModule to app module.
