+++ app/src/test/kotlin/com/jpaya/englishisfun/firestore/FireStorePropertiesTest.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2020 Jose Maria Payá Castillo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.jpaya.englishisfun.firestore
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+class FireStorePropertiesTest {
+
+    private val fireStoreProperties = FireStoreProperties()
+
+    @Test
+    fun verifyAbbreviationCollectionName() {
+        assertEquals(fireStoreProperties.getAbbreviationCollectionName(), "abbreviation")
+    }
+
+    @Test
+    fun verifyAbbreviationDocumentName() {
+        assertEquals(fireStoreProperties.getAbbreviationDocumentName(), "list")
+    }
+
+    @Test
+    fun verifyAbbreviationListField() {
+        assertEquals(fireStoreProperties.getAbbreviationListField(), "abbreviations")
+    }
+}
